# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "utoipa-axum",
    "description": "Utoipa's axum bindings for seamless integration for the two",
    "contact": {
      "name": "Juha Kukkonen",
      "email": "juha7kukkonen@gmail.com"
    },
    "license": {
      "name": "MIT OR Apache-2.0",
      "identifier": "MIT OR Apache-2.0"
    },
    "version": "0.2.0"
  },
  "paths": {
    "/projects": {
      "get": {
        "operationId": "list_projects",
        "responses": {
          "200": {
            "description": "Comments found successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_ProjectPreview"
                }
              }
            }
          }
        }
      }
    },
    "/projects/{id}": {
      "get": {
        "operationId": "get_project",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of project to fetch",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Project found successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Page_ProjectPreview": {
        "type": "object",
        "required": [
          "items",
          "has_previous",
          "has_next"
        ],
        "properties": {
          "has_next": {
            "type": "boolean"
          },
          "has_previous": {
            "type": "boolean"
          },
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "id",
                "name",
                "preview",
                "thumbnailId",
                "datePosted"
              ],
              "properties": {
                "datePosted": {
                  "type": "string",
                  "format": "date-time"
                },
                "id": {
                  "type": "string"
                },
                "name": {
                  "type": "string"
                },
                "preview": {
                  "type": "string"
                },
                "thumbnailId": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          }
        }
      },
      "Project": {
        "type": "object",
        "required": [
          "id",
          "name",
          "thumbnailId",
          "datePosted"
        ],
        "properties": {
          "datePosted": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "projectUrl": {
            "type": [
              "string",
              "null"
            ]
          },
          "thumbnailId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "ProjectPreview": {
        "type": "object",
        "required": [
          "id",
          "name",
          "preview",
          "thumbnailId",
          "datePosted"
        ],
        "properties": {
          "datePosted": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "preview": {
            "type": "string"
          },
          "thumbnailId": {
            "type": "string",
            "format": "uuid"
          }
        }
      }
    }
  }
}